+++ src/eckit/cmd/PsCmd.cc
@@ -177,12 +177,12 @@ void PsCmd::execute(std::istream& in, std::ostream& out, CmdArg& args) {
         grep = std::string(args["grep"]);
     }
 
-    for (size_t i = 1; i < args.size(); ++i)
+    for (long i = 1; i < args.size(); ++i)
         if (args.exists(i)) {
             all = false;
             if (args[i].isNumber()) {
                 int id = args[i];
-                if (info.size() >= id)
+                if (info.size() >= (unsigned long)id)
                     taskids.push_back(id);
                 else
                     pids.push_back(id);
