+++ src/tests/mpi/eckit_test_mpi.cc
@@ -68,7 +68,7 @@ BOOST_AUTO_TEST_CASE( test_all_to_all )
 
   // check results
   std::vector< std::vector<int> > expected(mpi::size());
-  for( int j=0; j<mpi::size(); ++j )
+  for(size_t j=0; j<mpi::size(); ++j)
     expected[j]=std::vector<int>(1,j);
 
   BOOST_CHECK_EQUAL_COLLECTIONS(recv.begin(),recv.end(),expected.begin(),expected.end());
@@ -87,7 +87,7 @@ BOOST_AUTO_TEST_CASE( test_all_gather___simple )
 
   // check results
   std::vector<int> expected(mpi::size());
-  for( int j=0; j<mpi::size(); ++j )
+  for(size_t j=0; j<mpi::size(); ++j)
     expected[j]=j;
 
   BOOST_CHECK_EQUAL_COLLECTIONS(recv.begin(),recv.end(),expected.begin(),expected.end());
@@ -106,9 +106,9 @@ BOOST_AUTO_TEST_CASE( test_all_gather___buffer )
 
   // check results
   std::vector<int> expected;
-  for( int j=0; j<mpi::size(); ++j )
+  for(size_t j=0; j<mpi::size(); ++j )
   {
-    for( int i=0; i<j; ++i )
+    for( size_t i=0; i<j; ++i )
       expected.push_back(j);
   }
